Modificando funções de experimento e finalizando experimento do python

+++ arq_publico.h
@@ -25,7 +25,7 @@ typedef struct Heap
 } Heap;
 
 typedef struct DadosCSV{ 
-    int iterações;
+    float iterações;
     int numDados;
 }DadosCSV;
